removed second controller and started POST request for generating schedule
